Combined hybrid module
