+++ src/main/webapp/site/src/app/teacher/teacher-routing.module.ts
@@ -39,7 +39,7 @@ const teacherRoutes: Routes = [
           }
         ]
       },
-      {path: '', loadChildren: () => import('./teacher-angular-js-module').then(m => m.TeacherAngularJSModule)}
+      {path: '', loadChildren: () => import('../hybrid-module').then(m => m.TeacherAngularJSModule)}
     ],
   }
 ];
